[Silver I] Title: 동물원, Time: 96 ms, Memory: 15032 KB -BaekjoonHub

+++ 백준/Silver/1309. 동물원/동물원.java
@@ -7,7 +7,7 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int n = Integer.parseInt(br.readLine());
-        long[][] dp = new long[n + 1][3];
+        int[][] dp = new int[n + 1][3];
         dp[1][0] = dp[1][1] = dp[1][2] = 1;
 
         for (int i = 2; i <= n; i++) {
